Remove public direct access to status flags
This is more in line with how the architecture should function, where
the CPU status is not directly exposed to programmers. Instead, the
status code should be interacted with via the appropriate Set and
Clear commands.

As a result, integration tests need to be updated; specifically, the
interrupt tests are now controlled via SEI and CLI commands.

This commit also fixes an oversight in the IRQ and NMI logic where we
were not correctly reading the interrupt program counter from memory.